The redirect for both buttons finally works #124 #125
